ftp/curl: Update to 8.15.0 and convert to CMake
Removes many old antics and makes it much more consistent with
others ports in tree. It also improve unit it testing as it
utilizes parallel processes. This also indirectly helps to
ensure that breakage doesn't occur in consumers such as CMake
which also uses the CMake build to build the internal version of
curl.